Apply changes and suggestions
